+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixes
+- Fix `Settings.initialize`on first run (create `.vscode` if not exists).
 
 ## 0.1.0 - 2019-10-05